perf: Divided Navbar into Navigation and NavigationMenu component
