+++ src/components/ui/navigation/Navigation.interfaces.ts
@@ -1,0 +1,3 @@
+export interface NavigationProps {
+    navigationNameData: string[];
+}
