+++ src/components/ui/navigationMenu/NavigationMenu.interfaces.ts
@@ -1,0 +1,3 @@
+export interface NavigationMenuProps {
+    navigationNameData: string[];
+}
